% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSdata.R
\name{sparse_dmDSdata}
\alias{sparse_dmDSdata}
\title{Create sparse dmDSdata object}
\usage{
sparse_dmDSdata(tx2gene, counts, samples)
}
\arguments{
\item{tx2gene}{Data frame which has to contain a \code{gene_id} column with
gene IDs, \code{feature_id} column with feature IDs.}

\item{counts}{matrix or sparseMatrix of count data, with a column for each sample. Rows correspond to features, for
example, transcripts or exons. Column names corresponding to
sample IDs must be the same as in the \code{sample} data frame.}

\item{samples}{Data frame where each row corresponds to one sample. Columns
have to contain unique sample IDs in \code{sample_id} variable and a
grouping variable \code{group}.}
}
\value{
Returns a \linkS4class{dmDSdata} object.
}
\description{
Convenience constructor function for a sparse \code{\linkS4class{dmDSdata}} object.
}
\seealso{
\code{\link{plotData}}
}
