% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sparseMatrixList.R
\docType{class}
\name{sparseMatrixList-class}
\alias{sparseMatrixList-class}
\alias{names,sparseMatrixList-method}
\alias{names<-,sparseMatrixList-method}
\alias{rownames,sparseMatrixList-method}
\alias{rownames<-,sparseMatrixList-method}
\alias{colnames,sparseMatrixList-method}
\alias{colnames<-,sparseMatrixList-method}
\alias{length,sparseMatrixList-method}
\alias{elementNROWS,sparseMatrixList-method}
\alias{dim,sparseMatrixList-method}
\alias{nrow,sparseMatrixList-method}
\alias{ncol,sparseMatrixList-method}
\alias{[[,sparseMatrixList-method}
\alias{$,sparseMatrixList-method}
\alias{[,sparseMatrixList,ANY,ANY,ANY-method}
\alias{[,sparseMatrixList-method}
\alias{[,sparseMatrixList,ANY-method}
\title{sparseMatrixList object}
\usage{
\S4method{names}{sparseMatrixList}(x)

\S4method{names}{sparseMatrixList}(x) <- value

\S4method{rownames}{sparseMatrixList}(x)

\S4method{rownames}{sparseMatrixList}(x) <- value

\S4method{colnames}{sparseMatrixList}(x)

\S4method{colnames}{sparseMatrixList}(x) <- value

\S4method{length}{sparseMatrixList}(x)

\S4method{elementNROWS}{sparseMatrixList}(x)

\S4method{dim}{sparseMatrixList}(x)

\S4method{nrow}{sparseMatrixList}(x)

\S4method{ncol}{sparseMatrixList}(x)

\S4method{[[}{sparseMatrixList}(x, i, j)

\S4method{$}{sparseMatrixList}(x, name)

\S4method{[}{sparseMatrixList,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{sparseMatrixList object.}

\item{value, i, j, name}{Parameters used for subsetting and assigning new
attributes to x.}
}
\value{
\itemize{ 
\item \code{names(x)}, \code{names(x) <- value}: Get or set names
of matrices. 
\item \code{rownames(x)}, \code{rownames(x) <- value},
\code{colnames(x)}, \code{colnames(x) <- value}: Get or set row names or
column names of unlistData slot. 
\item \code{length(x)}: Get the number of
matrices in a list. 
\item \code{elementNROWS(x)}: Get the number of rows of each of
the matrices. 
\item \code{dim(x)}, \code{nrow(x)}, \code{ncol(x)}: Get the
dimensions, number of rows or number of columns of unlistData slot. 
\item
\code{x[[i]]}, \code{x[[i, j]]}: Get the matrix i, and optionally, get only
columns j of this matrix. 
\item \code{x$name}: Shortcut for
\code{x[["name"]]}. 
\item \code{x[i, j]}: Get a subset of sparseMatrixList that
consists of matrices i with columns j. }
}
\description{
A sparseMatrixList object is a container for a list of matrices which have the same
number of columns but can have varying number of rows. Additionally, one can
store an extra information corresponding to each of the matrices in
\code{metadata} matrix.
}
\section{Slots}{

\describe{
\item{\code{unlistData}}{Matrix which is a row binding of all the matrices in a list.}

\item{\code{partitioning}}{List of indexes which defines the row partitioning of
unlistData matrix into the original matrices.}

\item{\code{metadata}}{Matrix of additional information where each row corresponds to
one of the matrices in a list.}
}}

